Section-5/Multi-Agent-4/requirements.txt
crewai
boto3
google-cloud-compute
google-cloud-billing
google-genai
pydantic
python-dotenv
crewai-tools